remove int cast

+++ src/estimator.py
@@ -34,7 +34,7 @@ def severe_cases_by_time(infections):
     return (infections*15/100.0)
 
 def hospital_beds_by_time(beds_count, infections):
-    available_beds = int(beds_count * 35/100.0)
+    available_beds = (beds_count * 35/100.0)
     return available_beds - infections
 
 def icu_request_by_time(infections):
